Data class back to double

+++ restaurant-ordering-system-api/src/main/java/api/Inventory/Item.java
@@ -15,10 +15,10 @@ public class Item {
 	//@Column(name = "description")
 	private String description;
 	//@Column(name = "price")
-	private int price;
+	private double price;
 	private boolean isAvailable;
 	
-	public Item(String name, String description, int price, boolean isAvailable) {
+	public Item(String name, String description, double price, boolean isAvailable) {
 		this.id = _id;
 		_id++;
 		this.name = name;
@@ -28,7 +28,7 @@ public class Item {
 		
 	}
 	
-	public Item(int id, String name, String description, int price, boolean isAvailable) {
+	public Item(int id, String name, String description, double price, boolean isAvailable) {
 		this.id = id;
 		_id=id+1;
 		this.name = name;
@@ -56,7 +56,7 @@ public class Item {
 	public void setDescription(String description) {
 		this.description = description;
 	}
-	public int getPrice() {
+	public double getPrice() {
 		return price;
 	}
 	public void setPrice(int price) {
